lavague_weather/requirements.txt
black
beautifulsoup4
flask
lavague
lavague-core
nltk
python-dotenv
requests
transformers
wikipedia

